Created new module unpacker and moved unpack_file_url() to it
